+++ src/Accounts/Accounts/Account/DisconnectAzureRmAccount.cs
@@ -16,7 +16,6 @@ using Microsoft.Azure.Commands.Common.Authentication;
 using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
 using Microsoft.Azure.Commands.Common.Authentication.Models;
 using Microsoft.Azure.Commands.Profile.Models;
-// TODO: Remove IfDef
 #if NETSTANDARD
 using Microsoft.Azure.Commands.Profile.Models.Core;
 #endif
@@ -127,6 +126,12 @@ namespace Microsoft.Azure.Commands.Profile
                     if (GetContextModificationScope() == ContextModificationScope.CurrentUser)
                     {
                         AzureSession.Instance.AuthenticationFactory.RemoveUser(azureAccount, environment: null);
+                    }
+                    else if ((string.Equals(azureAccount.Type, AzureAccount.AccountType.ServicePrincipal, StringComparison.OrdinalIgnoreCase)
+                        || string.Equals(azureAccount.Type, "ClientAssertion", StringComparison.OrdinalIgnoreCase))
+                        && AzureSession.Instance.TryGetComponent(PowerShellTokenCacheProvider.PowerShellTokenCacheProviderKey, out PowerShellTokenCacheProvider tokenCacheProvider))
+                    {
+                        tokenCacheProvider.ClearAppTokenCache();
                     }
 
                     if (AzureRmProfileProvider.Instance.Profile != null)